got curuthers in 2d
got it rendering a cat in all the right ways but its 2d :/

+++ src/GEPEngine/Environment.cpp
@@ -25,12 +25,12 @@ namespace GEPEngine
 		gettimeofday(&tv, NULL);
 		double curr = (tv.tv_sec) * 1000 + (tv.tv_usec) / 1000;
 #endif
-		float diff = curr - last;
+		double diff = curr - last;
 		deltaTime = diff / 1000.0f;
 		last = curr;
 	}
 
-	float Environment::getDT()
+	double Environment::getDT()
 	{
 		return deltaTime;
 	}
